Implement webpack into the ilearnx app
